refactor(progress,schedule): モジュール分割で保守性を向上
- progress/page.tsx: 1,214行 → 268行（9ファイルに分割）
- TodaySchedule.tsx: 786行 → 270行（4ファイルに分割）
- カスタムフック抽出: useWorkProgressActions, useTodayScheduleSync
- ユーティリティ関数外部化: progress/utils.ts

Closes #62
